+++ powerauth-backend-tests/src/test/java/com/wultra/security/powerauth/test/v40/PowerAuthActivationTest.java
@@ -140,9 +140,6 @@ class PowerAuthActivationTest {
         PowerAuthActivationShared.activationPrepareWithoutInitTest(config, model);
     }
 
-    // Test activationPrepareBadMasterPublicKeyTest is skipped, because master public key is not used
-    // in protocol version 4.0 for encryption due to use of temporary keys.
-
     @Test
     void activationStatusTest() throws Exception {
         PowerAuthActivationShared.activationStatusTest(powerAuthClient, config, model, VERSION);
